backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
python-socketio==5.11.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.2
bcrypt==4.2.0
marshmallow==3.21.3
requests==2.32.3